+++ src/main/java/vn/edu/hcmuaf/fit/entity/Province.java
@@ -3,24 +3,24 @@ package vn.edu.hcmuaf.fit.entity;
 import java.io.Serializable;
 
 public class Province implements Serializable {
-    private int id;
+    private long id;
     private String name;
     private String prefix;
 
     public Province() {
     }
 
-    public Province(int id, String name, String prefix) {
+    public Province(long id, String name, String prefix) {
         this.id = id;
         this.name = name;
         this.prefix = prefix;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
